canaal-flask/requirements.txt
absl-py==1.2.0
astunparse==1.6.3
attrdict==2.0.1
attrs==22.1.0
boto3==1.24.92
botocore==1.27.92
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
conda==4.3.16
cytoolz==0.11.2
dlib==19.24.0
docx2txt==0.8
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
flake8==4.0.1
flatbuffers==22.9.24
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.49.1
gunicorn==20.1.0
h5py==3.7.0
hexbytes==0.2.2
idna==3.3
importlib-metadata==5.0.0
iniconfig==1.1.1
jmespath==1.0.1
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
lru-dict==1.1.7
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.6.1
numpy==1.22.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
parsimonious==0.8.1
pluggy==1.0.0
protobuf==3.19.6
py==1.11.0
py-solc==3.2.0
py-solc-x==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pycodestyle==2.8.0
pycosat==0.6.3
pycryptodome==3.14.1
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
rlp==2.0.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
scipy==1.8.0
semantic-version==2.9.0
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
tensorflow-metal==0.6.0
termcolor==2.0.1
tomli==2.0.1
toolz==0.11.2
typing_extensions==4.3.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.9
web3==4.7.2
websockets==6.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
